+++ crates/edr_provider/src/error.rs
@@ -16,6 +16,7 @@ use edr_chain_spec::{
 use edr_chain_spec_block::BlockChainSpec;
 use edr_eth::{filter::SubscriptionType, BlockSpec, BlockTag};
 use edr_evm_spec::{result::ExecutionResult, DatabaseComponentError, TransactionError};
+use edr_gas_report::GasReportCreationError;
 use edr_primitives::{hex, Address, Bytes, B256, U256};
 use edr_rpc_eth::{client::RpcClientError, error::HttpError, jsonrpc};
 use edr_runtime::{
@@ -28,8 +29,8 @@ use edr_state_api::StateError;
 use serde::Serialize;
 
 use crate::{
-    config::IntervalConfigConversionError, debug_trace::DebugTraceError,
-    gas_reports::GasReportCreationError, time::TimeSinceEpoch, ProviderSpec,
+    config::IntervalConfigConversionError, debug_trace::DebugTraceError, time::TimeSinceEpoch,
+    ProviderSpec,
 };
 
 /// Helper type for a chain-specific [`CreationError`].