+++ crates/edr_rpc_eth/tests/integration/client.rs
@@ -466,7 +466,14 @@ mod alchemy {
             .await
             .expect_err("should have failed");
 
-        assert!(matches!(error, RpcClientError::JsonRpcError { .. }));
+        let RpcClientError::HttpStatus(error) = error else {
+            panic!("Expected HttpStatus error, got: {error:?}");
+        };
+
+        assert_eq!(
+            reqwest::Error::from(error).status(),
+            Some(reqwest::StatusCode::BAD_REQUEST)
+        );
     }
 
     #[tokio::test]
@@ -583,7 +590,14 @@ mod alchemy {
             .await
             .expect_err("should have failed");
 
-        assert!(matches!(error, RpcClientError::JsonRpcError { .. }));
+        let RpcClientError::HttpStatus(error) = error else {
+            panic!("Expected HttpStatus error, got: {error:?}");
+        };
+
+        assert_eq!(
+            reqwest::Error::from(error).status(),
+            Some(reqwest::StatusCode::BAD_REQUEST)
+        );
     }
 
     #[tokio::test]