+++ crates/edr_solidity_tests/tests/it/helpers/config.rs
@@ -131,7 +131,7 @@ impl<
 
     /// Executes the test runner
     pub async fn test(self) -> BTreeMap<String, SuiteResult<HaltReasonT>> {
-        self.runner.test_collect(self.filter).await
+        self.runner.test_collect(self.filter).await.suite_results
     }
 
     pub async fn run(self) {